backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
opencv-python==4.9.0.80
insightface==0.7.3
onnxruntime==1.17.0
numpy==1.26.4
pyyaml==6.0.1
pydantic==2.6.1
python-multipart==0.0.9
